move func test to base
